flytt og separer sykmelding typer til 'types' mappe
