+++ src/types/sykmeldingCommon.ts
@@ -1,0 +1,26 @@
+export enum Blad {
+    A = 'A',
+    B = 'B',
+}
+
+export enum LottOgHyre {
+    BEGGE = 'BEGGE',
+    HYRE = 'HYRE',
+    LOTT = 'LOTT',
+}
+
+export enum ArbeidssituasjonType {
+    ANNET = 'ANNET',
+    ARBEIDSLEDIG = 'ARBEIDSLEDIG',
+    ARBEIDSTAKER = 'ARBEIDSTAKER',
+    FISKER = 'FISKER',
+    FRILANSER = 'FRILANSER',
+    JORDBRUKER = 'JORDBRUKER',
+    NAERINGSDRIVENDE = 'NAERINGSDRIVENDE',
+    PERMITTERT = 'PERMITTERT',
+}
+
+export type FomTom = {
+    readonly fom: string
+    readonly tom: string
+}
